Fix game list ordering
Game ordering was broken when clicking on the name column or
adding/removing a favorite.

Unlike the standard lib's that requires a strict weak ordering
comparator, wxWidgets needs a compare function that returns a negative
number, 0 or a positive number when entries are respectively lower than,
equal to and greater than.

Fixed using C++20 comparison API.