Add bundle type extensions to CMakeLists.txt
Adds MACOSX_BUNDLE_TYPE_EXTENSIONS to CMakeLists, and sets `wua` as a claimed filetype. 

(It doesn't have to be called this, as CMake only handles a few Info.plist properties. I only called it this to fit with the existing ones. Personally I think it shouldn't be called MACOSX since the OS is up to 14 and left 10 behind a long time ago).