UI: wxCAPTION flag on input API dialog to fix kwin
The IP and port input boxes are not shown when the DSUController API is
first selected when using wayland with kwin.

Setting the dialog as wxCAPTION causes kwin to add decorations to the
window which for whatever reason causes the toplevel to resize
correctly.

This may be a bug in wxWidgets or kwin.