Revert use of new, slower wxListCtrl::Set*Images API in game list
The SetImageList gives ListCtrl a pointer to use for the images, which we can then asynchronously load its items from. There's no locking, but before we upgraded it to use the new API it has never really caused any issues. In 40302a706b1387fa293fa18f2a9f51111f87d455 I upgraded to the recommended API that's mentioned in the wxWidgets docs.

However, the performance, especially with a lot of game titles installed, was quite terrible with this new API due to us needing to call wxListCtrl::Set*Images each time a new icon was loaded, which it'll have to rescale each time for the user's DPI (at least, it seems to spend a long time in a bitmap/icon rescaling method). https://github.com/cemu-project/Cemu/issues/1672 brought this up. So ultimately I think it's just better from a usability standpoint to use the old methods.

Also, I removed some of the hacks that were used for the icon column width size, now that its using the new upstreamed fix. I also added a flag to draw vertical lines between the columns.